backend/wata/requirements.txt
httpx>=0.24.0
psycopg2-binary>=2.9.0
